feat(estoque): etiqueta sem preço + recalc automático de balanço
Duas mudanças alinhadas com o André sobre gestão de preço médio:

1. ETIQUETA SEM PREÇO
   - Remove "CUSTO: R$ X" das duas variantes de etiqueta (regular + pendência/troca)
   - Etiqueta agora mostra só: QR code + modelo+cor + serial/IMEI + fornecedor
   - Preço fica vinculado ao serial no banco de dados — vendedor consulta via scan
   - Nunca mais precisa reimprimir etiqueta por mudança de preço

2. RECALC AUTOMÁTICO DE BALANÇO
   - Extrai lógica de recalcBalancos para lib/recalc-balancos.ts (função compartilhada)
   - Chama automaticamente após: importação em lote e inserção individual de produto
   - Quando chega lote novo com preço diferente, o custo_unitario (balanço) de TODAS as
     unidades do mesmo modelo é recalculado como média ponderada
   - Endpoint /api/admin/recalc-balancos simplificado para reusar a mesma lib
   - Ex: 2 un a R$8.000 + 2 un a R$7.500 → balanço = R$7.750 pra todas as 4
